oss commit
source update

+++ src/5geir/EquipmentIdentityCheck/handler/api_equipment_status_document.go
@@ -31,7 +31,7 @@ func GetEquipmentStatus(c *gin.Context) {
 
 	// Check Mandatory Parameter
 	if request.Query["pei"] == nil {
-		status := http.StatusBadRequest
+		status := int32(http.StatusBadRequest)
 		detail := scenario.ErrorDetailMandatoryIeIncorrect
 		cause := scenario.MandatoryIeIncorrect
 		response := model.ProblemDetails{Status: &status, Detail: &detail, Cause: &cause}
